Version 2
Made the system that draws the text on the bottom half better, and added a line width slider.
Reduced the delay for adding the image from .3s to .1s.
Made it so that if you have a wide screen it'll put the controls and card side by side instead of forcing vertical.
Moved copyright to the top of the screen.
Made centering the caption work.